Add files via upload
This is the first iteration of the model with rough predictions. I plan on exploring the features more and refining the selection as well as addressing the severe class imbalance. Subsequently, I also cull the suite of models (if necessary) and hypertune the parameters to improve the specificity of the model. Code for SVM also needs to be updated with relevant caret tuning functions.